# Some perf tweaks