--- a/Cargo.toml
+++ b/Cargo.toml
@@ -54,6 +54,7 @@ peeking_take_while = "0.1.2"
 quote = { version = "0.6", default-features = false }
 regex = "1.0"
 which = "2.0"
+fxhash = "0.2"
 # New validation in 0.3.6 breaks bindgen-integration:
 # https://github.com/alexcrichton/proc-macro2/commit/489c642.
 proc-macro2 = { version = "0.4", default-features = false }