--- a/src/lib.rs
+++ b/src/lib.rs
@@ -23,7 +23,7 @@ extern crate cexpr;
 #[allow(unused_extern_crates)]
 extern crate cfg_if;
 extern crate clang_sys;
-extern crate fxhash;
+extern crate hashbrown;
 #[macro_use]
 extern crate lazy_static;
 extern crate peeking_take_while;
@@ -97,9 +97,9 @@ use std::process::{Command, Stdio};
 use std::sync::Arc;
 
 // Some convenient typedefs for a fast hash map and hash set.
-type HashMap<K, V> = ::fxhash::FxHashMap<K, V>;
-type HashSet<K> = ::fxhash::FxHashSet<K>;
-pub(crate) use ::std::collections::hash_map::Entry;
+type HashMap<K, V> = ::hashbrown::HashMap<K, V>;
+type HashSet<K> = ::hashbrown::HashSet<K>;
+pub(crate) use ::hashbrown::hash_map::Entry;
 
 fn args_are_cpp(clang_args: &[String]) -> bool {
     return clang_args