I did a bit of profiling today, for no good reason. I wanted to get rid of some of our hash maps from before hand so it's good to know that this really helps.

```
$ time ./bindgen-old tests/stylo.hpp --no-rustfmt-bindings >/dev/null 2>&1                                                                                                                           
./bindgen-old tests/stylo.hpp --no-rustfmt-bindings > /dev/null 2>  6.06s user 0.76s system 98% cpu 6.912 total
$ time ./bindgen-new tests/stylo.hpp --no-rustfmt-bindings >/dev/null 2>&1                                                                                                                                      
./bindgen-new tests/stylo.hpp --no-rustfmt-bindings > /dev/null 2>&1  4.51s user 0.63s system 98% cpu 5.198 total
```