@emilio 2018-12-29
Well, I'll merge this for now, since it's a pretty nice improvement and I want to pull it on mozilla-central. But if you have time to skim over it it'd be also really amazing :)